Upgraded to Laravel 12

+++ config/api/app/config.php
@@ -6,5 +6,5 @@ return [
     'admin_email' => env('ADMIN_EMAIL'),
     'registrations' => env('APP_REGISTRATIONS', false),
     'internal_api_key' => env('INTERNAL_API_KEY'),
-    'token_expiry_days' => env('API_TOKEN_EXPIRY_DAYS', 90),
+    'token_expiry_days' => (int) env('API_TOKEN_EXPIRY_DAYS', 90),
 ];
